+++ packages/angular-workspace/spright-angular/chat/input/testing/input.pageobject.ts
@@ -1,0 +1,3 @@
+import { ChatInputPageObject } from '@ni/spright-components/dist/esm/chat/input/testing/chat-input.pageobject';
+
+export { ChatInputPageObject };