+++ packages/storybook/src/spright/chat/input/chat-input-matrix.stories.ts
@@ -1,0 +1,73 @@
+import type { StoryFn, Meta } from '@storybook/html';
+import { html, ViewTemplate } from '@ni/fast-element';
+import {
+    bodyFont,
+    bodyFontColor
+} from '@ni/nimble-components/dist/esm/theme-provider/design-tokens';
+import { chatInputTag } from '@ni/spright-components/dist/esm/chat/input';
+import {
+    sharedMatrixParameters,
+    createMatrixThemeStory,
+    createMatrix
+} from '../../../utilities/matrix';
+import { createStory } from '../../../utilities/storybook';
+import { hiddenWrapper } from '../../../utilities/hidden';
+import { textCustomizationWrapper } from '../../../utilities/text-customization';
+import { loremIpsum } from '../../../utilities/lorem-ipsum';
+
+const valueStates = [
+    ['empty', ''],
+    ['one line', "My cat's breath smells like cat food."],
+    ['multi line', `${loremIpsum} ${loremIpsum} ${loremIpsum}`]
+] as const;
+type ValueState = (typeof valueStates)[number];
+
+const placeholderStates = [
+    ['empty', ''],
+    ['one line', 'This is the placeholder']
+] as const;
+type PlaceholderState = (typeof placeholderStates)[number];
+
+const metadata: Meta = {
+    title: 'Tests Spright/Chat Input',
+    parameters: {
+        ...sharedMatrixParameters()
+    }
+};
+
+export default metadata;
+
+// prettier-ignore
+const component = (
+    [valueLabel, value]: ValueState,
+    [placeholderLabel, placeholder]: PlaceholderState
+): ViewTemplate => html`
+    <p 
+        style="
+        font: var(${bodyFont.cssCustomProperty});
+        color: var(${bodyFontColor.cssCustomProperty});
+        margin-bottom: 0px;
+        "
+    >    
+        ${valueLabel} value, ${placeholderLabel} placeholder
+    </p>
+    <${chatInputTag}
+        placeholder="${placeholder}"
+        value="${value}"
+    >
+    </${chatInputTag}>
+`;
+
+export const themeMatrix: StoryFn = createMatrixThemeStory(
+    createMatrix(component, [valueStates, placeholderStates])
+);
+
+export const hidden: StoryFn = createStory(
+    hiddenWrapper(
+        html`<${chatInputTag} hidden>Hidden Chat Input</${chatInputTag}>`
+    )
+);
+
+export const textCustomized: StoryFn = createMatrixThemeStory(
+    textCustomizationWrapper(html`<${chatInputTag}></${chatInputTag}>`)
+);